requirements.txt
Django==1.6.4
MySQL-python==1.2.3
PySimpleSOAP==1.10
PyYAML==3.11
South==0.8.4
Unidecode==0.04.16
beautifulsoup4==4.3.2
ebaysdk==2.0.0
flake8==2.1.0
gevent==1.0.1
greenlet==0.4.2
grequests==0.2.0
lxml==3.3.5
python-slugify==0.0.7
requests==2.2.1
wsgiref==0.1.2
